@@ -1,58 +1,62 @@
 {
-    "name": "ecomengine",
-    "version": "0.0.0",
-    "scripts": {
-        "404": "copy dist\\index.html dist\\404.html",
-        "ng": "ng",
-        "start": "ng serve --port 4200",
-        "build": "ng build --prod",
-        "lint": "ng lint"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^7.2.0",
-        "@angular/common": "^7.2.0",
-        "@angular/compiler": "^7.2.0",
-        "@angular/core": "^7.2.0",
-        "@angular/forms": "^7.2.0",
-        "@angular/platform-browser": "^7.2.0",
-        "@angular/platform-browser-dynamic": "^7.2.0",
-        "@angular/router": "^7.2.0",
-        "animate.css": "^3.7.0",
-        "bootstrap": "^4.2.1",
-        "core-js": "^2.6.2",
-        "dexie": "^2.0.4",
-        "font-awesome": "^4.7.0",
-        "jquery": "^3.3.1",
-        "moment": "^2.23.0",
-        "popper.js": "^1.14.6",
-        "primeicons": "^1.0.0",
-        "primeng": "^7.0.4",
-        "roboto-fontface": "^0.10.0",
-        "rxjs": "^6.3.3",
-        "tslib": "^1.9.3",
-        "zone.js": "^0.8.27"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^0.12.1",
-        "@angular/cli": "^7.2.1",
-        "@angular/compiler-cli": "^7.2.0",
-        "@angular/language-service": "^7.2.0",
-        "@types/jasmine": "^3.3.5",
-        "@types/jasminewd2": "^2.0.6",
-        "@types/node": "^10.12.18",
-        "codelyzer": "^4.5.0",
-        "jasmine-core": "^3.3.0",
-        "jasmine-spec-reporter": "^4.2.1",
-        "karma": "^3.1.4",
-        "karma-chrome-launcher": "^2.2.0",
-        "karma-coverage-istanbul-reporter": "^2.0.4",
-        "karma-jasmine": "^2.0.1",
-        "karma-jasmine-html-reporter": "^1.4.0",
-        "protractor": "^5.4.2",
-        "ts-helpers": "^1.1.2",
-        "ts-node": "^7.0.1",
-        "tslint": "^5.12.1",
-        "typescript": "^3.2.2"
-    }
+  "name": "ecomengine",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --port 4200",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint --type-check",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~8.2.11",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~8.2.11",
+    "@angular/compiler": "~8.2.11",
+    "@angular/core": "~8.2.11",
+    "@angular/forms": "~8.2.11",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.11",
+    "@angular/platform-browser-dynamic": "~8.2.11",
+    "@angular/router": "~8.2.11",
+    "ajv": "^6.9.1",
+    "animate.css": "^3.7.2",
+    "bootstrap": "^4.4.1",
+    "core-js": "^3.4.5",
+    "dexie": "^2.0.4",
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.3.1",
+    "moment": "^2.23.0",
+    "popper.js": "^1.16.0",
+    "primeicons": "^1.0.0",
+    "primeng": "^7.0.3",
+    "roboto-fontface": "^0.10.0",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.12",
+    "@angular/cli": "~8.3.12",
+    "@angular/compiler-cli": "~8.2.11",
+    "@angular/language-service": "~8.2.11",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-helpers": "^1.1.2",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
+  }
 }
